Closes #11094

_Briefly describe your proposed changes:_

This moves the WAL from being in the tsm1.Engine to being in the storage.Engine. This required a number of changes including:

- Some weird upward signaling and lock management
- Changing retention to use DeleteBucketRange
- Removing some delete methods from the engine
- Removing some delete type from the WAL
- Adding a new delete type to the WAL
- Replay the WAL rather than directly load into the Cache
- Move the tsm1 Value type into its own package