@zeebo 2019-02-04
### Added TODOs

- https://github.com/influxdata/influxdb/pull/11364/files#diff-18b0fccd77bafb3a93fcacdfdd86ec34R228
	This existed before this PR. This TODO just calls out the poor behavior.

- https://github.com/influxdata/influxdb/pull/11364/files#diff-b44a6551f36995f10c237a575eec2bf7R558
	This TODO is just moving the lock usage one into the function body where it happens and adding a comment about keeping track of the keys. I should probably remove the key comment because it's actually just a slice header per key being removed, since the cache already keeps all of the keys in memory.

- https://github.com/influxdata/influxdb/pull/11364/files#diff-b633e12e59f6e30d9f2f96ccc8a301e2R108
	I'd like some help in knowing which metrics to provide here. How does adding timing + status of every delete for every bucket/org combo sound? Worried about cardinality? Just a histogram of timing + counts of errors/ok?

- https://github.com/influxdata/influxdb/pull/11364/files#diff-18b0fccd77bafb3a93fcacdfdd86ec34R420
	I'm waiting for ben's change for improving performance of explode so that I can make this work well with that.

### Action items

- Update comment in the cache.
- Add metrics when which ones are agreed upon.