@zeebo 2019-02-04
### Performance

    inch -bucket b -org o -p 10 -m 10 -t 100,100,10 -v -c 8

#### Master
```
10000000 points written (100%).
Total throughput: 689477.0 pt/sec | 689477.0 val/sec.
Current throughput: 350000 val/sec.
Errors: 0
μ: 114.874573ms
90%: 150.491507ms
95%: 173.744145ms
99%: 233.363966ms
```

#### This change

```
10000000 points written (100%).
Total throughput: 693866.6 pt/sec | 693866.6 val/sec.
Current throughput: 330000 val/sec.
Errors: 0
μ: 114.452816ms
90%: 156.055782ms
95%: 176.183093ms
99%: 220.710342ms
```

### Conclusion

There is no performance change. We're doing the same amount of work, just in a different order.